+++ packages/rocketh-deploy-proxy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/deploy-proxy
 
+## 0.10.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/deploy@0.10.6
+
 ## 0.10.5
 
 ### Patch Changes